requirements.txt
Flask==1.1.2
RPi.GPIO==0.7.0
websockets==9.0
